% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_smoking}
\alias{lives_smoking}
\title{Percentage of Current Smokers (2022-2023)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{smoking_percentage}{Percentage of people that are current smokers}
\item{year}{Time period}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://www.health-ni.gov.uk/publications/health-survey-northern-ireland-first-results-202223}
}
\usage{
lives_smoking
}
\description{
A dataset containing statistics on smoking status in
each Northern Irish Council, 2022-23.
}
\keyword{datasets}
