% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_all_mortality}
\alias{people_all_mortality}
\title{Mortality all causes rate per 100k (2020-22)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{death_rate_per_100k}{age-sex standardised rate for all causes of
mortality per 100k}
\item{year}{3-year aggregated period}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
people_all_mortality
}
\description{
A dataset containing statistics from an age-sex standardised rate for all
causes of mortality per 100k, by Council (2020-22).
}
\details{
To note: England's Health Index uses a single year to calculate mortality rates
from all causes. Northern Ireland only has 3-year aggregate data available.
}
\keyword{datasets}
