% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_avoidable_deaths}
\alias{people_avoidable_deaths}
\title{Rate of Avoidable Deaths per 100k (2018-2022)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{avoidable_deaths_per_100k}{Age-standardised avoidable mortality rates per
100k}
\item{year}{5-year aggregate period}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
people_avoidable_deaths
}
\description{
A dataset containing statistics on age-standardised death rates per 100k, by
Northern Irish Council (2018-2022).
}
\keyword{datasets}
