% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_cancer}
\alias{people_cancer}
\title{Cancer Registrations Percentage (2023-2024)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{cancer_percentage}{Percentage of people with cancer per Council}
\item{year}{Time period - Financial year}
...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
people_cancer
}
\description{
A dataset containing statistics on cancer registration percentages, per
Council (2023-2024).
}
\keyword{datasets}
