% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_cardiovascular_conditions}
\alias{people_cardiovascular_conditions}
\title{Percentage Rate of Cardiovascular Conditions (CHD, Atrial Fibrillation,
Heart Failure, Stroke & TIA) (2023/2024)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authroity Code}
\item{cardiovascular_conditions_percentage}{Percentage of people who self-reported
as having CHD, Atrial Fibrillation, Heart Failure, and/or Stroke & TIA}
\item{year}{Time period - financial year}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
people_cardiovascular_conditions
}
\description{
A dataset containing statistics on the percentage of people who self-reported
as having CHD, Atrial Fibrillation, Heart Failure, and/or Stroke & TIA in
Northern Irish Council Areas.
}
\keyword{datasets}
