% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_dementia}
\alias{people_dementia}
\title{Percentage of People with Dementia}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{dementia_percentage}{Percentage of people medically diagnosed with
dementia, per Council}
\item{year}{Time period - Financial year}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
people_dementia
}
\description{
A dataset containing statistics on the percentage of people medically
diagnosed with dementia, by Council (2023-24).
}
\keyword{datasets}
