% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_diabetes}
\alias{people_diabetes}
\title{Percentage prevalence of diabetes (2023-2024)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{diabetes_percentage}{Percentage prevalence of diabetes (all types)}
\item{year}{Time period}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
people_diabetes
}
\description{
A dataset containing statistics on the percentage prevalence of people with
diabetes, by Northern Irish Council (2023-2024).
}
\keyword{datasets}
