% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_disability_daily_activities}
\alias{people_disability_daily_activities}
\title{Percentage of People Aged 15-64 with Long-Term Health Problem/Disability
that Limits Daily Activities}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{disability_activities_limited_percentage}{percentage of people aged 15-64
with long-term health problem/disability that limits daily activities by a
lot or by a little.}
\item{year}{Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://www.nisra.gov.uk/publications/census-2021-main-statistics-health-disability-and-unpaid-care-tables}
}
\usage{
people_disability_daily_activities
}
\description{
A dataset containing statistics on the percentage of people aged 15-64 with
long-term health problem/disability that limits daily activities, by Council
(2021).
}
\keyword{datasets}
