% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_life_expectancy}
\alias{people_life_expectancy}
\title{Average  Life Expectancy for Men and Women (2020-2022)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{life_expectancy_combined}{Average life expectancy for men and women}
\item{year}{Time period - three year aggregate}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/healthandlifeexpectancies/datasets/lifeexpectancyforlocalareasinenglandnorthernirelandandwalesbetween2001to2003and2020to2022}
}
\usage{
people_life_expectancy
}
\description{
A dataset containing statistics on average life expectancy for men
and women, by Northern Irish Council (2020-2022).
}
\keyword{datasets}
