% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_household_overcrowding}
\alias{places_household_overcrowding}
\title{Percentage Household Overcrowding (2017)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{household_overcrowding_percentage}{Percentage of households
experiencing household overcrowding}
\item{year}{Year}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
places_household_overcrowding
}
\description{
A dataset containing statistics on the percentage of households experiencing
household overcrowding, per Council (2017).
Household overcrowding is typically defined as households with an occupancy
rating of -1 or lower.
}
\keyword{datasets}
