% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_personal_crime}
\alias{places_personal_crime}
\title{Rates of Personal Crime per 1,000 (2017)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{personal_crime_per_1k}{Number of personal crimes per 1,000 people}
\item{year}{Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
places_personal_crime
}
\description{
A dataset containing statistics on 'personal crime' per 1,000 people in
Northern Irish Council Areas. 'Personal crime' is defined as violence against
the person, sexual offences, robbery, theft, criminal damage, and arson.
}
\keyword{datasets}
