% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_sports_centre_travel_time_sdz}
\alias{places_sports_centre_travel_time_sdz}
\title{Travel time to sports facilities}
\format{
A data frame with 11 rows and 7 variables:
\describe{
\item{sdz21_code}{Super Data Zone Code}
\item{sports_centre_mean_travel_time}{Mean travel time (in minutes) to the nearest
sports centre. 999 is a pseudo-value that indicates there are no sports
centres within 3 hours of the SDZ.}
\item{is_within_3_hours}{Is there a sports centre within 3 hours of the centre of
the SDZ?}
\item{year}{Year the data was last updated}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}
...
}
}
\source{
\url{https://openstreetmap.org/}
}
\usage{
places_sports_centre_travel_time_sdz
}
\description{
A dataset containing the average (median) travel time (in minutes) to the nearest
sports centre in each Super Data Zone.
}
\details{
Travel times are calculated using the TravelTime API (https://traveltime.com/apis/distance-matrix)
and are based on travelling by public transport on a weekday morning.
}
\keyword{datasets}
