% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_unemployment}
\alias{places_unemployment}
\title{Percentage of People Experiencing Unemployment (2023)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{unemployment_percentage}{Percentage of people experiencing unemployment
and claiming unemployed-related benefits(Jobseeker’s Allowance plus
out-of-work Universal Credit claimants who were claiming principally
for the reason of being unemployed.)}
\item{year}{Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
places_unemployment
}
\description{
A dataset containing statistics on the percentage of people experiencing
unemployment, claiming unemployed-related benefits.
}
\keyword{datasets}
