% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_alcohol_misuse}
\alias{lives_alcohol_misuse}
\title{Rate of Patient Admissions for Alcohol Related Conditions (2019)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{alcohol_admissions_rate_per_100k}{Rate of patient admissions per 100k
population, based on population numbers per HSCT, converted to Local
Authority level. Patients are counted only once in the year in which they
have an alcohol-related stay.}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.health-ni.gov.uk/publications/census-drug-and-alcohol-treatment-services-northern-ireland-30th-april-2019}

\url{https://data.nisra.gov.uk/}
}
\usage{
lives_alcohol_misuse
}
\description{
A dataset containing statistics on the rate of patient admissions for
all alcohol-related admissions in each Northern Irish Council, 2019.
}
\keyword{datasets}
