% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_gsce_attainment}
\alias{lives_gsce_attainment}
\title{Percentage GSCE Attainment, including Maths and English (2022/23)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{gcse_qualifications_percent}{Grade A-C percentage of GSCEs, including
Maths and English on results day, by Council, end of school year 2023}
\item{year}{School Year}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
lives_gsce_attainment
}
\description{
A dataset containing statistics on the percentage of GSCE attainment,
including Maths and English (A-C grade).
}
\keyword{datasets}
