% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_healthy_eating}
\alias{lives_healthy_eating}
\title{Percentage of People that Eat Healthy (2022-2023)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{healthy_eating_percentage}{Percentage of people that consume
5 portions or more of fruit and vegetable per day}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.health-ni.gov.uk/publications/health-survey-northern-ireland-first-results-202223}
}
\usage{
lives_healthy_eating
}
\description{
A dataset containing statistics on fruit and vegetable consumption in
each Northern Irish Council, 2022-23.
}
\keyword{datasets}
