% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_high_blood_pressure}
\alias{lives_high_blood_pressure}
\title{Percentage of Adults Diagnosed with High Blood Pressure (2023-2024)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{high_blood_pressure_percentage}{Percentage of adults diagnosed with
high blood pressure}
\item{year}{Time period - Financial year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
lives_high_blood_pressure
}
\description{
A dataset containing statistics on percentage of adults diagnosed with high
blood pressure, by Council (2023-2024).
}
\keyword{datasets}
