% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_low_birth_weight}
\alias{lives_low_birth_weight}
\title{Percentage of Babies Born Not at a Healthy Birth Weight (2022-2023)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{low_birth_weight_percentage}{Percentage of babies born not at a healthy birth weight}
\item{year}{Time period}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://www.publichealth.hscni.net/publications/statistical-profile-childrens-health-northern-ireland-202223}
}
\usage{
lives_low_birth_weight
}
\description{
A dataset containing percentage babies born not at a healthy birth weight.
}
\keyword{datasets}
