% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_pupil_absence}
\alias{lives_pupil_absence}
\title{Percentage of Pupil Absences (2021/22)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{total_pupil_absence_percentage}{Percentage of unauthorised school
absences across primary, secondary, and special schools}
\item{year}{School Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
lives_pupil_absence
}
\description{
A dataset containing statistics on the percentage of unauthorised school
absences.
}
\keyword{datasets}
