% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_respiratory_conditions}
\alias{people_respiratory_conditions}
\title{Percentage Rate of Respiratory Conditions (Asthma and COPD) (2023/2024)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authroity Code}
\item{respiratory_conditions_percentage}{Percentage of people who self-reported
as having asthma and/or COPD}
\item{year}{Time period - financial year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
people_respiratory_conditions
}
\description{
A dataset containing statistics on the percentage of people who self-reported
as having asthma and/or COPD in Northern Irish Council Areas.
}
\keyword{datasets}
