% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_internet_access}
\alias{places_internet_access}
\title{Percentage of Households without Internet Access (2021/22)}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{no_internet_access_percentage}{Percentage of households without access
to the internet}
\item{year}{Financial Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://data.nisra.gov.uk/}
}
\usage{
places_internet_access
}
\description{
A dataset containing statistics on the percentage of households without
internet access in Northern Irish Council Areas.
}
\keyword{datasets}
