% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_noise_complaints}
\alias{places_noise_complaints}
\title{Rates of Noise Complaints per 1,000 (2023-2024)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{noise_complaints_per_1k}{Number of noise complaints per 1,000 people}
\item{year}{Year}

...
}
}
\source{
https://www.daera-ni.gov.uk/sites/default/files/publications/daera/Noise\%20Complaint\%20Statistics\%20for\%20NI\%202023-24.PDF
}
\usage{
places_noise_complaints
}
\description{
A dataset containing statistics on noise complaints per 1,000 people in
Northern Irish Council Areas.
}
\keyword{datasets}
